refactor(auth): Move security into its own module (#645)
LDAP tests are still in web since they need the `Main` class.
