+++ gate-core/src/main/groovy/com/netflix/spinnaker/gate/filters/CorsFilter.groovy
