+++ gate-core/src/main/groovy/com/netflix/spinnaker/gate/services/internal/Front50Service.groovy
