+++ gate-security/gate-security.gradle
@@ -1,0 +1,39 @@
+apply from: "$rootDir/gradle/ldap.gradle"
+
+repositories {
+  maven {
+    url 'https://repo.spring.io/libs-snapshot'
+  }
+}
+
+dependencies {
+  spinnaker.group('fiat')
+
+  compile project(':gate-core')
+  compile spinnaker.dependency("kork")
+  // RetrySupport is in kork-exceptions and not kork-core TODO: Move to Core
+  compile spinnaker.dependency("korkExceptions")
+  compileOnly spinnaker.dependency("lombok")
+
+  // OAuth2
+  compile 'org.springframework.security.oauth:spring-security-oauth2:2.1.1.RELEASE'
+  compile 'com.squareup.retrofit:converter-simplexml:1.9.0' // Github OAuth
+
+  // LDAP
+  compile("org.springframework.security:spring-security-ldap:${spinnaker.version('springSecurity')}")
+
+  // SAML
+  compile('org.opensaml:opensaml:2.6.4')
+  compile("org.springframework.security.extensions:spring-security-saml2-core:1.0.2.RELEASE")
+  compile("org.springframework.security.extensions:spring-security-saml-dsl:1.0.0.M3") {
+    exclude group: 'org.springframework.boot'
+    exclude group: 'org.springframework.security'
+  }
+
+  // IAP
+  compile('com.nimbusds:nimbus-jose-jwt:5.2')
+
+  spinnaker.group('test')
+  testCompile apachedsDependencies
+  testCompile spinnaker.dependency("korkJedisTest")
+}
