+++ gate-security/src/main/groovy/com/netflix/spinnaker/gate/security/SpinnakerAuthConfig.groovy
