+++ gate-security/src/main/groovy/com/netflix/spinnaker/gate/security/anonymous/AnonymousConfig.groovy
