+++ gate-security/src/main/groovy/com/netflix/spinnaker/gate/security/oauth2/OAuthSsoConfigurer.groovy
