+++ gate-security/src/main/groovy/com/netflix/spinnaker/gate/security/x509/X509UserIdentifierExtractor.java
