+++ gate-security/src/test/groovy/com/netflix/spinnaker/gate/security/oauth2/ExternalAuthTokenFilterSpec.groovy
