+++ gate-security/src/test/groovy/com/netflix/spinnaker/gate/security/x509/OidRolesExtractorSpec.groovy
