+++ gate-security/src/test/resources/memberOfNoRoles.crt
