+++ gate-security/src/test/resources/memberOfTwoRoles.crt
